Rewrite project to use SteamKit2/C#
The existing ValvePython/steam and ValvePython/csgo libraries have
ceased to function. I've decided that it makes the most sense to rewrite
on top of the canonical library for interacting with Steam.

Just like the previous version of this package was my introduction to
Python and the original version was my introduction to PHP, this version
is my introduction to C# and _vibe_ coding. We'll see how this actually
ends up working out.